Fixed problem with gmeta exports. Fixed urls for UBC data.
